+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.5.0] - 2023-06-30
+
 ### Added:
 
 - Support for `GET /api/v1/b/:bucket/:entry/batch` endpoint, [PR-62](https://github.com/reductstore/reduct-js/pull/62) 
@@ -154,7 +156,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release
 
-[Unreleased]: https://github.com/reduct-storage/reduct-js/compare/v1.4.1...HEAD
+[Unreleased]: https://github.com/reduct-storage/reduct-js/compare/v1.5.0...HEAD
+
+[1.5.0]: https://github.com/reduct-storage/reduct-js/compare/v1.4.1...v1.5.0
 
 [1.4.1]: https://github.com/reduct-storage/reduct-js/compare/v1.4.0...v1.4.1